Qt5.8 QDateTime: Update the housekeeping code.
This is mostly a straight substitution of calls to ::fromTime_t with
calls to ::fromSecsSinceEpoch.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).