Qt5.8 QDateTime: Update protocol handling for START_RECORDING subcommand.
Update the handling for the QUERY_REMOTEENCODER/START_RECORDING
command.  This also updates the QUERY_COMMBREAK, QUERY_CUTLIST,
QUERY_BOOKMARK, and SET_BOOKMARK protocol command handlers (which
aren't called anywhere in MythTV).  These are all straight
substitution of calls to ::{to,from}Time_t with calls to
::{to,from}SecsSinceEpoch.  All references are to a recording start
time, so are all assumed to be valid and no extra checking for
validity was added.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).