Qt5.8 QDateTime: Update InfoMap for a ProgramInfo instance.
This is a straight substitution of calls to ::toTime_t with calls to
::toSecsSinceEpoch with change of variables from int to qint64.  The
start/end timestamps for a recording are assumed to be valid, so no
extra validation was added.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).