Qt5.8 QDateTime: Update the "set watched" code.
This is a straight substitution of calls to ::toTime_t with calls to
::toSecsSinceEpoch, and then some code cleanup to make the code easier
to understand.

The QDateTime {to,from}Time_t functions were obsoleted in Qt5.8 and
replaced by {to,from}SecsSinceEpoch functions.  The two sets of
functions perform the same operation, only the latter set takes a
wider range of values than the former (uint vs. qint64).