Frontend GUI: Add support for long presses of selected keys
Long press will work for Shield remote and for keyboards,
for selected keys only.

I added a new section to "edit keys" called "Long Press".
You can assign up to 16 keys that will support log presses.
Initially only "Enter" is in there.

I added a default setting for Menu in the global settings,
assigned to a long press of Enter. This only takes effect
for new frontends. To use it with an existing frontend,
use "edit keys" to add a long press of enter to Global Menu.

Adding a key to "Long Press" will disable autorepeat for the key.

Refs #13234