Frontend GUI: Menu and Bookmark enhancements
Added SETBOOKMARK and TOGGLEBOOKMARK actions in TV Playback context.
On a new frontend they are not assigned to any key.

If an existing frontend gets upgraded to the new version, the keys
assigned to SELECT in the Global context will be assigned to either
SETBOOKMARK or TOGGLEBOOKMARK depending on whether the "alternate set
and remove bookmark" setting was checked.

The "alternate set and remove bookmark" setting is removed from frontend
setup and is no longer used.

On a new frontend, the space bar will be added as a key on the Pause
action. This does not happen on an existing frontend being upgraded.

Set bookmark, Toggle bookmark and Jump to bookmark are added to the
Playback navigation menu.

The Program details screen that normally only switches between the two
layouts with INFO will also switch with SELECT action (e.g. ENTER)

Video list page: switch from the group menu to the file menu when MENU
is pressed a second time. This is similar to what the Recordings list does.

Refs #13234