Clang: Eliminate illegal character encoding in character literal warnings.
Clang complains about the arguments to the QLatin1Char function being
"high-ascii" values which aren't valid in a source file using the
utf-8 encoding.  Change these characters to hex values to eliminate
the warnings.  The object files produced from the clang compiler from
before and after this change was compared to make sure they are
identical.