Clang: Eliminate illegal character encoding in character literal warnings.
Clang complains about characters in strings that use "high-ascii"
values, because these aren't valid in a source file using the utf-8
encoding.  Change these characters to hex values to eliminate the
warnings.  The lines are very long and not very pretty.  This in order
to ensure that the resulting file has the exact same number of lines.
This make it possible to compare the object files produced from the
clang compiler from before and after this change was compared to make
sure they are identical.