Fix GCC8 warning about cases falling through.
There is no possibility of fallthrough, as the switch() is being
performed on an enum and all values of the enum have been accounted
before.  It seems like the problem is that the variable is of type int
instead of enum, but adding a static cast doesn't eliminate the
warning messages.  Add explicit break statements to make the compiler
happy, even though they serve no other purpose.