Fix Clang6 about mheg using two different enums in a switch statement.
The MHEG code defines two different enums with essentially the same
names and values.  In the CallProgram function, in a single switch
statement it uses one value from one enum, and the remaining values
from the other enum.  Convert the function to be consistent and only
use the values from a single enum.  The .o files from before/after
this change were compared and there was no difference in the compiled
code.