Remove the clang "-Wno-self-assign" compiler flag.
Also cleans up the one location in the code that does self assign.
This is an operator= function for copying an object, and the existing
comment says it is intentionally keeping the old values.  Commenting
out the self-assignment (instead of removing it) eliminates the error,
but also highlights which fields are kept from the old object.