Fix GCC8 warning about a 'case value not in enumerated type'.
The code is using "QEvent::User" and "QEvent::User + 1" to denote two
different events.  This is part of the number space allocated by Qt
for user defined events, but only the first actually appears in the Qt
enum.  The latter, and any other user defined events used with the
QEvent::Type type, will cause compiler warnings.  Solve this by
disabling the GCC "-Wswitch" warning for this one and only instance of
the problem.