GUI: Enhance Playback to allow change of recording group with arrow keys
New feature to help with limited remotes such as on the
NVidia Shield. On the playback screen, allow the arrow keys
to navigate to the recording groups buttons and select a new
recording group. This is disabled unless the user selects
new option "Change Recording Group using the arrow keys" in
setup.
This requires some things to be done in themes:
(recordings-ui.xml).
1. add focusorder 2 to recgroups buttonlist
2. add state selectedactive and active to recgroups buttonlist
   to highlight when the recgroups have focus.
3. Remove triggerevent left from groups buttonlist to allow "left"
   to get to recgroups or set wrapstyle "none" to allow "up" to get
   to recgroups.
These have been done in the mythcenter and mythcenter wide themes.

Refs #13234