Playback OSD: New navigation screen
When invoked, OSD Navigation provides a row of icons along the bottom of the
playback screen with items like play, pause, jump forward, jump back.
There is a "more" button that shows other selections of icons.
The navigation icons are displayed in conjunction with the osd_status
(progress bar).

There is a new keyboard action available called OSDNAVIGATION.
This needs to be assigned to a key in setup under TV Playback.
I recommend assigning "Enter", so that the enter button can be used
on the NVidia Shield.

There is a default OSD Navigation window included in the default and defaultwide
themes, that will show on all themes unless overridden.

Theme designers can create as many buttons as they
wish and assign them to as many groups as they wish. They must be set
up so as not to overlap with the osd_status display. Actions
that are available in TV Playback can be added to buttons as desired
by the theme designer. There is no fixed list of actions, they can
be decided by the theme designer.

I will document it fully in the wiki for theme designers.

Refs #13234