Fix crash (abort) in MythSocket::ResetReal
In libmythbase mythsocket.cpp MythSocket::ResetReal there can be cases
where there are no bytes to available to read, so the vector is never
resized to greater than zero, which results in an abort when
referencing the 1st (non-existant) element which is now checked by
gcc8 (libstdc++).

Patch from Gary Buhrmaster, fixes #13264.