Change shebang from '/usr/bin/env perl' to /usr/bin/perl.
Use of /usr/bin/env has a storied discussion, but is considered ill
advised by some. For all the perl code in the main sources the shebang
specifies /usr/bin/perl with the exception of the following two
files. Some build processes report a warning of the use of
/usr/bin/env (and/or modify it to meet the package standards).

Patch from Gary Buhrmaster, fixes #13272.