WebFrontend: Fix commercial flag icon and background-positions
From the dev list, thanks Philipp.

For commerical flagged recordings show the "flag" icon instead of the
"clock" icon, which is for auto expiry.

All 'background-position's must be given as *negative* values, as it
describes the start position relative to the top-left corner of the HTML
element where painting the background images starts.

Signed-off-by: Bill Meek <bmeek@mythtv.org>