Services API: Adds Dvr/ManageJobQueue endpoint
Allows clients to queue Commflag, Metadata,
Transcode and UserJob<x>.

Dvr/ManageJobQueue

* Action Add:
  requires: JobName, RecordedId
  optional: JobStartTime, default=current time
  optional: JobArgs, default=none (jobqueue.args)
  optional: RemoteHost, defaults to GetHostName()

The jobqueue.comment column wil be set to:
Dvr/ManageJobQueue.

Returns: -1 for errors or the new JobId

* Action Remove:
  requires: JobName, RecordedId, JobId

Returns: -1 for errors or the JobId.

Myth/ManageDigestUser and MythManageUrlProtection
are recommended prior to using this (and other
endpoints that POST for that matter.)