Setup: Support typed input into any spinbox
Allows for typed entry into any spinbox in the system, including those
on the front end. Pressing enter or a number button or key into a
spinbox opens a dialog where you can type any number that is valid
for the spinbox. If a number is not valid for the spinbox
the "OK" button is disabled.

Fixes #13204

(cherry picked from commit 862e510f589e3fedb8f37b61ac7967a219b7d8f4)