Fix how configure finds libxml2.
Remove the first component of the include file name as it is redundant
with the path provided by pkg-config.  This is not normally noticable
as on most systems the default include path contains the parent of the
libxml include path, and it all works out.  On FreeBSD this is not the
case, and the libxml2 libraries can't be found without tweaking.