Use the system provided libXNVCtrl on all Linux systems.
This is an interim step to using the system provided libXNVCtrl on all
systems.  The FreeBSD package for nvidia-settings currently builds
this library as libXNVCtrl.a and then links it into the
nvidia-settings application.  A request has been made to the package
maintainer to build libXNVCtrl.so and install that alongside the
nvidia-settings application.