Clean up bluray includes in bdringbuffer.cpp.
The libbluray header files include two identical sets of values for
video rates, where the only difference is the name prefix.  Switching
from the BD_ prefix to the BLURAY_ prefix allows the removal of a
couple of include files, and will allow easier removal of the internal
copy of libbluray at some point in the future.