Use the system provided libbluray (when >= 0.9.3).
The copy of libbluray embedded into MythTV was recently updated and is
now current, but it would be better to use the system provided
libbluray where possible.  MythTV apparently requires a minimum
version of 0.9.3, so only use the system libbluray if its a later
version.  Eventually once all the distributions with old versions of
libbluray become unsupported, the embedded copy of libbluray can be
removed from the MythTV sources.