Fix a couple more EIT issues.
Reset the next scan time after a tuning failure.  This causes scanning
to resume after a reasonable time instead of 1 year as previously
done.

Start on a random channel the first time scanning is done.  After
that, resume on the next channel.  This ensures that channels
immediately after a channel the fails to tune get scanned in a timely
manner.

Fixes #12106
Fixes #13269