Allow recording of the full, unfiltered, MPTS from a transport.
Add a "Add full Transport Stream channel" option to the channel
scanner. then the channel scanner will create a 'MPTS' channel for each
transport scanned.

If the recorder is told to record that 'MPTS' channel, then all of the PIDs
in the transport stream will be recorded. No filtering will be done.

In parallel, individual programs from the same transport can be recorded as
well.