Frontend Setup: Fix shift key handling in Edit Keys
Detection of key codes in Edit Keys was inconsistent with the way they are checked
in the main window. Shift is ignored for cetain key ranges in order to allow
alphabetic commands to be case insensitive. This check was not done in
Edit Keys. The result was there were certain keys it was not possible
to use once assigned, for example "@" or Ctrl-Shift-B.

Fixes #13279