MythGUI: Change spinbox entry dialog to only appear for numeric spin boxes
The popup entry dialog is really only intended for entry of numeric values
when there is a huge range. The popup with alphabetic entry is interfering
with custom record. It was not useful anyway so will be disabled.

Fixes #13318

(cherry picked from commit dc0755bbd656cedff4a0af21919306ef3019e593)