C++11 modernization: Use nullptr. (libs/libmythui/mythrender_opengl2.cpp)
The last argument to the m_glVertexAttribPointer function is declared
as a pointer, but is passed either a pointer or an integer depending
upon the value of the first argument.  Add a new inline function that
hides the integer to char* cast from the compiler, and call that in
cases where the last argument is an integer.  This is an inline, so it
shouldn't add any instructions to the code execution path.