VAAPI Rewrite: Support the latest VAAPI from FFmpeg.
The new vaapi code is called vaapi2 in the profile selection.
It adds the following:
- Fixes problems where vaapi fails when there is a resolution change
  during playback.
- Adds support for vaapi hardware deinterlacing.
- Fixes vaapi broken under wayland.
- Adds support for decoder based deinterlacing. Previously deinterlacing was
  always part of rendering. Now both methods are possible.

It uses the standard OpenGL rendering in the code, unlike the prior vaapi
support, which created a special render class called OpenGLVAAPI.

It does not share code with the old vaapi. The old vaapi can be disabled
and eventually removed.

Fixes #13311