A couple of tweaks to configure.
1) Fix the blank line before the "XXX CONFIGURATION" lines.  Its a
problem because the behavior of the the echo command differs between
the default Fedora and Ubuntu shells. Switching to /bin/echo doesn't
work because the FreeBSD version of the command doesn't support the -e
flag.  The simplest solution is just an extra echo command.  (Extra
credit goes to the person who fixes the configure script to the bash
shell so its consistent across all platforms.)

2) The errlog function was removed from the configure script in 2016.
Fix up the last caller to this function.