Support database wakeup without needing external wakeonlan program
When deploying on android we cannot simply install the wakeonlan program
and reference it in the database setup. Since there is already code in
MythTV to do a wake on lan for waking slave backends, change  database
wakeup to use the same code. Simply putting the MAC in the wakeup
command will perform a wake on lan.