C++11 modernization: Remove deprecated header files. (plugins)
A number of C library header files were deprecated in C++, and their
C++ equivalents should be used instead.  Now that a C++11 compiler is
required these replacements can be made.  While working on the
headers, sort the includes in the areas that are already being
modified.  These header changes were both suggested by the clang-tidy