DVD playback: Fix deadlock
A deadlock during playback of DVDs happens in recent master.
This is likely a result of two-levels of queuing that is
now supported for decoders that produce multiple frames
from a single packet. This change removes a lock that seems
unnecessary as it is only accessing and not updating.

Fixes #13326