TV Playback: Fix playback of Music Choice channels
Fix audio interruptions in playing music choice channels.
There is a new setting to enable Music Choice. This needs to be
selected for uninterrupted music choice playback. Whan this is enabled,
there is a small chance that the code may misidentify regular video as
music choice and that would cause lip sync problems. If that happens,
Music choice should be disabled in settings.

Fixes #13294