OpenGL Video Playback: Add settings to disable pixel formats
The frontend already has the ability to disable some pixel formats via
an environment string. This adds those options as settings is a new Video
Playback Advanced section. This is needed for android as it is
difficult to pass environment strings to an android program and
for best performance we need to disable YV12 pixel format on some
android devices. Also revert the earlier changes that always disabled
it for android.