Playback: Improved Video timing and Synchronization (avsync2)
Since modern videos have audio and video timestamps, it makes more
sense to use the timestamp to display each frame at the required time
rather than assuming a frame rate that may be wrong. A new synchronization
routine (AVSync2) works this way. It uses a real time clock via a
system function and shows each frame at the calculated correct time.
The frame is actually shown at the next refresh time, since the renderers
use a dual buffering scheme. AVSync2 adjusts audio synchronization in a
different way, by changing the video time base. There is a setting that
can be adjusted to improve audio sync when the audio card does not provide
the application with good data on the amount of sound buffered.

With this patch, the old sync routine is still the default. To use AVSync2,
you need to select it in frontend settings.

Fixes #13334