Playback: Avsync2: Fix for when player shows one frame for a long time
If decoding is slow, the player drops frames to catch up. However
if the decoding is slower than playback, it never catches up, leaving
a single frame shown. The fix limits the number of sequential
dropped frames and also pauses audio so that video can catch up
in extreme situations

Refs #13334