C++11 modernization: Use default member initialization.
The C++11 standard allows specifying default values for member
variables as part of the class definition.  This makes specifying them
again in a constructor as redundant, and allows the code to be cleaned
up and easier to read.  It may also allow constructors to be specified
as =default allowing the compiler to make further optimizations.

These changes were suggested by the clang-tidy program.