Fully support HDHomeRun tuner locking and pooling.
Previously, with HDHomeRuns, MythTV inputs were configured to use
specific tuners.  This didn't work well when other applications were
configured to use the same HDHomeRuns.  If the specific tuner was busy
when MythTV needed it, MythTV would fail.  With this change, each
MythTV input is now configured to use one or more HDHomeRuns.  When a
tuner is needed, MythTV will use the first free tuner from the
configured set of HDHomeRuns.

Note that multiple MythTV inputs must still be configured in order to
use multiple HDHomeRun tuners simultaneously.  For example, if an
HDHomerun Quatro has 4 tuners, 4 MythTV inputs should be configured.
If two HDHomeRun Primes have 6 tuners total (3 + 3), 6 MythTV inputs
should be configured.

If other applications might be using tuners from the same HDHomeRuns,
it is up to the user to make sure the appropriate number of tuners are
available to MythTV.  If MythTV needs a tuner and none is available,
MythTV will still fail.  One way to prevent that is to configure fewer
MythTV inputs than there are HDHomerun tuners available.  For example,
for an HDHomeRun Quatro, only configure 2 MythTV inputs leaving 2
inputs available to other applications.  Alternatively, set the
recording priorities of some inputs lower so that MythTV will not try
to use them except when absolutely necessary.

Finally, this change contains a database schema update to
automatically update the HDHomeRun input configuration for optimum
use.  It is commented out at this time in order to not break
compatibility with MythTV version 29.  It will be uncommented before
the release of version 30.  In the meantime, the configuration can be
updated simply by running mythtv-setup, entering the Capture Card
section, immediately exiting and saving the changes.  Until the
configuration is updated, MythTV will continue to function as before
with each MythTV input still using the specific HDHomeRun tuner it was
previously configured to use.

Fixes #13339