Python Bindings: Releases MythTV.services_api module
This module can be used in Python programs to access
Services API endpoints. It can also serve as a menu
of tasks that a user may want to include (inline)
in their programs.

This module works with Python versions 2.7 and 3.n,
but read on.

For complete information, use Python's help feature:

    $ python
    >>> from MythTV.services_api import send as s, utilities as u
    >>> help(s)
    >>> help(u)

Sample import lines:

    import MythTV.services_api.send as api
    import MythTV.services_api.utilities as util

For Python3, the following builds and installs OK:

    cd ... mythtv/mythtv/bindings/python
    2to3-2.7 --write .
    cd ../..
    ./configure --python=python3 ...

But importing MythTV.services_api will fail. Seems,
2to3 doesn't care for everything, or does too much.

You can make this module available to Python 3.n (as
long as the above 2 to 3 conversion wasn't done.) For
example:

    sudo mkdir /usr/local/lib/python3.6/dist-packages/MythTV
    sudo ln --symbolic \
        /usr/local/lib/python2.7/dist-packages/MythTV/services_api \
        /usr/local/lib/python3.6/dist-packages/MythTV/services_api

Users of the original mythtv_services_api module that
was installed via pip should remove it/them. Same for
private copies in ~/bin etc.

    sudo --set-home pip2 uninstall --yes mythtv_services_api
    sudo --set-home pip3 uninstall --yes mythtv_services_api