Eliminate "useless cast" warnings related to QStrings. (part two)
Running GCC with the "useless-cast" warning enabled produces a number
of warnings.  Clean up places where multiple QStrings and C strings
are added together to create a QString, and then arguments are applied
to that QString.  Also use qPrintable a couple of places (instead of
the longer incantation) to make the code easier to read.