Fix Kernel, OpenGL Deinterlacer.
This change reverts the openglvideo part of commit 4aa24550.
According to Mark Kendall <mark_kendall@btinternet.com>, the original
author of the OpenGL code for MythtV, "It's just wrong".  In my
testing, this indeed does fix the kernel deinterlacer, which should,
once again, be better than the linear blend deinterlacer.