Fix shift overflow error in string_to_guid on 32bit platforms.
The old version of this function produced different values on 32bit
and 64bit platforms.  On a 32bit platform, the shift ended up being a
no-op, so the high and low halves of the number were folded together
producing a 32bit number.  The simple solution is to replace all this
code with a pair of Qt function calls.