Improve initial channel selection when starting live TV.
First, use the last viewed channel information for the frontend
instead of for the input that happens to be selected.  The channel
shouldn't necessarily change just because the input did.

Second, use the callsign and channum for the last viewed channel
instead of the chanid.  This is similar to how the EPG chooses the
channel when starting live TV.

Third, fallback to any viewable channel if none with the desired
callsign and channum are available.

Things to do after the release of v30 due to risk of breakage.

Remove legacy uses of channel.startchan and DefaultChanId setting.

Better honor capturecard.livetvorder.  It is only partially honored in
this commit.