Cache FileSystemInfo data to avoid unwanted blocking.
Change the MainServer to cache updated, FileSystemInfo data and
optionally return the cached data when requested.  The Scheduler uses
this data to decide where to put a recording right before it starts.
Previously, an unresponsive, remote filesystem could block the
Scheduler for up to several minutes.  Now, the Scheduler always uses
the cached data to avoid getting blocked.  The AutoExpirer does not
and should not use the cached data so that the data gets updated
periodically.