MythRenderOpenGL: Use QOpenGLTexture
- N.B. this removes support for YCbCr video texture types - they are not
supported by QOpenGLTexture
- they weren't widely supported anyway, not at all with GLES and they
are equivalent to the default opengl/UYVY code (which also supports full
colourspace control).

- this is a fairly invasive conversion but should greatly improve
portability and on some systems performance.
- it is slightly complicated by poor compatibility between
QOpenGLFramebufferObject and QOpenGLTexture which means we still need to
do wrapping of the framebuffer's texture and raw OpenGL for setting
texture filters and wrap for that texture.

- sundry cleanup of MythRenderOpenGL