Restore (cached) call to GetFilesystemInfos() in AutoExpire.
A previous change removed it because it was believed to be redundant.
It actually is redundant in that it doesn't need to be called again so
soon.  However, a logic bug left the needed filesystem information
empty.  Rather than attempt to thread safely re-use the data completely
within AutoExpire, restore the call to GetFilesystemInfos() but used
the previously cached results.

(cherry picked from commit 193eb990ceb30d6cd69c57972be7f9d380d1368e)