cppcheck: Clean up "variable assigned a value that is never used" messages.
Mostly removing the variable assignment.  In one case the value is
assigned from a function call, one of the other parameters is a return
value and is used, so just quiet the warning.  In another case the
variable is only used inside of an ifdef, so declare the variable in
an ifdef.