AVSync2: Use frame timestamps to determine elapsed play time
MythTV uses frame count to determine the elapsed time of a recording
for display, jumps, etc. This can be inaccurate with variable rate
content and also with frame doubling being done in the decoder.
This change recalculates the frame count based on the time code.
This ensures MythTV elapsed time will be accurate.

Fixes #13416