Valgrind error in DVBStreamHandler / DeviceReadBuffer
The DeviceReadBuffer is now stopped only once at the end of recording
instead of three times. One of the two later stop attempts accessed memory
that was already free'd in a previous stop. This caused valgrind errors
when testing with valgrind and this caused also possibly occasional
segmentation faults of  mythbackend at the end of a recording.

Fixes #13414