Video aspect ratio wrong on systems with multiple monitors
On a system with two monitors, positioned horizontally next to each other,
the vertical size of the video in a window is correct but the horizontal size
is about half what it should be.
The same is true for full-screen video, without an enclosing window.

This problem is most likely introduced in commit b2c82a3, dated Feb 1,
"Convert from QDesktopWindow to QScreen".

Before this commit a system with multiple monitors was treaded as "Xinerama".
After this commit a system with mulitple monitors is seen as one big screen
with an area which is the bounding box of the screens of all monitors.
The name display as used in the code now refers to the big "bounding box" screen.

The video aspect ratio computations use the size of the display in mm.
The bug was here that the display size in mm was used in conjunction
with the size in pixels of the first screen.
In order to scale correct the size in pixels of the display has to be used.

Note: the change from QDesktopWidget to QScreen is a Good Thing because
the QDesktopWidget is declared obsolete.
Note: with this fix included the video scaling in a multiple monitor
system is now exact which is better than before, in the "Xinerama" mode.

Fixes #13400