Update guide title/subtitle/description with new DVB EIT data
The title/subtitle/description in the guide is now updated
with the latest information received from DVB EIT. The title
is now updated with the newly received data unless the new
title is an empty string.
Previously, the title was only updated if the new title was
longer than the old title. This caused missed recordings when
recording rules are used that refer to the program title.
The same for the subtitle and the description.

Fixes #13057