No Signal/Noise and no Signal Strength with Si2168-based tuners
The Linux driver for the Si2168 only implements the DVBv5 calls
to retrieve the signal strength etc.
The MythTV code first tries the DVBv3 call and if that fails
with an EOPNOTSUPP then the DVBv5 call is tried.
This worked correct until Linux kernel changed the
error code from EOPNOTSUPP to ENOTSUPP (value 524).
This fix adds a test on ENOTSUPP in addition to the existing
test on EOPNOTSUPP.