MythPlayer: Refactor handling of EAGAIN when decoding video
- mediacodec direct rendering was stalling/deadlocking/failing in
numerous places as the decoder thread would get stuck continually
retrying for the next video frame when the underlying decoder is waiting
for a new bufer to render to.
- we cannot safely signal to the decoder thread that it needs to stop
retrying as we cannot guarantee getting a lock on the decoder.
- instead, move responsibility for handling EAGAIN into mythplayer,
which can try again if necessary or abort if the player is trying to
pause or exit.
- no obvious problems in testing...