Add exception for Digitenne (NL) network 8720 to enable EIT by default for all channels.
This exception enables EIT for all channels even when the flags in the SDT
tell that there is no EIT present. Analysis of a full transport stream shows
that the EIT information is present; with the fix applied the
guide information does appear in MythTV program guide.
This is definitely a problem in the Digitenne signal.
This problem was introduced at approximately 11 march 2019 when the modulation
system was changed from DVB-T to DVB-T2 and the video coding from H264 to HEVC.
When the Digitenne signal is correct again this exception can be removed.

Fixes #13427