Uninitialized memory access in video codec probes.
Valgrind detects uninitialized memory access in a memory buffer; the access is inside the buffer itself but beyond the size to which the buffer is written.
This access is done by FFmpeg video type probing routines. This could be viewed as a bug in these routines but the MythTV code does reserve 32 bytes of additional space for this so it is most likely a requirement for using these routines.  However, these 32 bytes are never initialized. This is solved by doing a "memset" to clear those 32 bytes.

Fixes #13428