Existing multiplexes not recognized with new DVB-T2 "Full Scan".
Doing a "Full Scan" with mythtv-setup for the first time the
channels are added and have the correct channel number.
The second run of "Full Scan" finds the same channels again
but sees them as new channels and they get the service ID number
because the correct channel number is already in use.
The third run gives conflicts on insertion because both
the correct channelnumber and the service id are already in use.

This problem is caused by the implementation of the
preliminary DVB-T2 support that has been added some time ago.
This creates some confusion whether a multiplex uses
the DVB-T or the DVB-T2 modulation system.
The strict comparison on the modulation system in
DTVMultiplex::IsEqual even for the "fuzzy" type  of
comparison makes that a new multiplex is not recognized
as already existing.

The solution for now is to treat the DVB-T and DVB-T modulation
systems as compatible when the "fuzzy" type of comparison is done.

Tickets #13014 and #12638 will comprise a more complete
implementation of DVB-T2 support.

Fixes #13432