tidy: Fix misleading indentation in mainserver.cpp.
The clang-tidy "misleading indentation" check pointed out a spot where
an 'if' statement and it opening/closing braces are conditionally
compiled, but the block of code inside the body of the 'if' is always
present.  Pulling the open/close braces outside of the conditional
compilation so that the compiler always sees them eliminates this
warning.

https://clang.llvm.org/extra/clang-tidy/checks/readability-misleading-indentation.html