tidy: Don't need to test a bool against 0/1 in an if statement.
The clang-tidy "implicit boolean conversion" check pointed out a
couple of places where a boolean value was being explicitly compared
to the number 0/1.  This requires an implicit conversion to an integer
to perform the check, and then an implicit conversion back to boolean
to determine whether to take the if/else branch.  Using the
pre-existing boolean value eliminates both conversions.

https://clang.llvm.org/extra/clang-tidy/checks/readability-implicit-bool-conversion.html